backend/test-requirements.txt
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.2.1
freezegun==1.2.2
httpx==0.24.1
coverage==7.3.0
django-webtest==1.9.10
model-bakery==1.20.5